pyproject.toml
[project]
name = "gm-curation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.7",
    "tweepy>=4.16.0",
]
